i2c multimaster: fix racy stop detection - RTC #49516
Driver wrongly assumes the interface always owns the bus when populating Tx
fifo: therefore it looks for a STOP symbol on the bus to terminate the
transaction.
This breaks in a multi-master scenario: if another master owns the bus while
the driver is filling the fifo, the subsequent STOP symbol on the bus will be
related to the other master.  This gets interpreted by the driver as its own
exit condition, resulting in a premature termination of the driver's
transaction.

Impact:
- data corruption: less bytes than expected are transmitted/received
- data corruption happens silently
- if the hardware is configured with IC_EMPTYFIFO_HOLD_MASTER_EN=1, the last
  byte which asserts the STOP command will not be transferred, resulting in a
  bus deadlock where SCL is held low.

The fix ensures STOP condition is processed only after all the content of the
transaction buffer is pushed onto Tx fifo, and both Tx and Rx fifos are empty.

Signed-off-by: Tan, Raymond <raymond.tan@intel.com>