Force a USB disconnect prior to USB enumeration in pch_udc driver to improve connection (rtc 80286)
Some Galileo and other reference boards show problems when connecting
as USB device. Host sees failure to enumerate and sticks faultily in
full speed. Device fails to enumerate correctly in High Speed and enters
a pattern of interrupt assertion for suspend/idle.

Manual workaround showed that reload of g_serial caused problem to go away.
By combining Elec Validation traces with probing of IRQ register a fault
pattern emerged, where device was in state with Idle and Suspend bits always
set. Acking of these bits would not clear them. This information was used as a
detection method, i.e. if we see IRQ register "stuck" with these bits we
can use it as an indication of the fault condition. It was also noted
that these IRQ bits were marked off by default after entering SUS so that
the fault condition could not be seen.

In a normal operation, the IRQ bits clear immediately after Acking and the
Suspend/Idle condition would be out of spec after going into suspend.

A threshold of 30 times was used to indicate that we have the fault
and then the workaround is implemented, which is to do a USB soft disconnect
from the device driver. This causes the device to re-enumerate with the
Host PC.

This re-enumeration probably clears the condition which caused
the problem also.

A standoff timer phy_err_time was implemented to only allow the recovery
attempt by default once every 100ms, module param allows htis to be changed
at module load. Workaround can be enabled by setting module param
phy_err_cnt to 30 (works well on Quark)  or other non zero value, smaller
values fix the problem earlier and are recommended.

Change-Id: I536080dde801d3c9467f26593a04765268256ecc
Signed-off-by: Kevin Clarke <kevin1.clarke@intel.com>
(cherry picked from commit ac0987ba8b431a962db113bda8114b33f602069c)

Signed-off-by: Dheeraj Jamwal <dheerajx.s.jamwal@intel.com>
Change-Id: I083d111d2c1b5440919889c9005e9a7da36926b5
Reviewed-on: https://git-gar-1.devtools.intel.com/gerrit/4781
Reviewed-by: Tan, Raymond <raymond.tan@intel.com>
Tested-by: Tan, Raymond <raymond.tan@intel.com>