tty: serial: sc16is7xx: Fix handling of RX FIFO overrun
If an RX FIFO overrun occurs, the IRQ handler for the
SC16IS7xx device will never return because the overrun
condition is not properly cleared.

In addition, the overrun is unlikely to be flagged to the
UART driver core in most cases because of an assumption in
the driver that bit 7 of the LSR register would be asserted
when the overrun error occurs, but this is not the case.

This patch adds a check for overrun on bit 1 of the LSR
register if overrun is suspected, and resets the RX FIFO
if the overrun condition is confirmed.

Signed-off-by: Dan O'Donovan <dan@emutex.com>
Signed-off-by: Kartikey Parmar <kartikeyx.r.parmar@intel.com>
Change-Id: Ibffb2917c895df027c3b417800f813b187f8f06d
Reviewed-on: https://git-gar-1.devtools.intel.com/gerrit/4647
Reviewed-by: Fung, Joshua <joshua.fung@intel.com>
Tested-by: Fung, Joshua <joshua.fung@intel.com>