tty: serial: sc16is7xx: Fix inverted gpio output on direction change
When changing the direction of a GPIO pin from input to output,
the initial logic level on the GPIO output may be incorrect.

It appears that the setting of the output state does not take effect
unless the pin direction is actually set to output first.

This patch switches the order in which the IODir and IOState register
are updated when changing the GPIO pin direction to output.

Alas, it is not possible to set the initial state of the GPIO output
immediately when the direction is set to output - some delay will occur.

Signed-off-by: Dan O'Donovan <dan@emutex.com>
Signed-off-by: Kartikey Parmar <kartikeyx.r.parmar@intel.com>
Change-Id: I907cc54d64cb6ef474b529aba8ed8adad2782292
Reviewed-on: https://git-gar-1.devtools.intel.com/gerrit/4648
Reviewed-by: Fung, Joshua <joshua.fung@intel.com>
Tested-by: Fung, Joshua <joshua.fung@intel.com>