Galileo Gen1 and Gen2 Board Platform Board File
This patch is the merging of a few files from BSP 1.0.1.

The platform board file of Galileo Gen1 and Gen2 is combined,
and the board is determined in realtime, which results in
different init() and exit() is called.

Besides, this patch contains multiple headers for different
platform peripherals that is available on both Gen1 and Gen2 platform.

Signed-off-by: Tan, Raymond <raymond.tan@intel.com>

platform: x86: intel-quark: update bus number for spi devices

In Quark, the bus number is fixed to 0 and 1, while the correct behavior
shall be read from the BUS ID. To preserve the behavior from QSD Development,
we made this patch for now.

Signed-off-by: Tan, Raymond <raymond.tan@intel.com>