gpiolib: undo retval changes of gpiochip_remove
This patch is added to undo the gpiochip_remove retval which was
not available during the development of Quark FFRD drivers based on
kernel 3.8.

However, while enabling the latest GPIO Designware Driver, which is
based on kernel 3.18, this gpiochip_remove function signature has
been changed to return type void.

Thus, there's this need to temporary revert the retval to int that
was introduced after the development of some Quark FFRD Drivers.

Signed-off-by: Tan, Raymond <raymond.tan@intel.com>