mfd: intel_qrk_gip: Enable GIP Driver for Intel X1000 SoC
This patch is to enable the GIP for Quark X1000 SoC. It was
part of a larger patch from Quark BSP R1.1 and it was adapted
to kernel version 3.14.

The explicit_stop logic of the original is removed, due to there
have been changes made in mainline to restart the i2c transaction.

The original author of this driver is
Josef Ahmad <josef.ahmad@linux.intel.com>

Signed-off-by: Tan, Raymond <raymond.tan@intel.com>