gpio-sch: RTC #96478 Fix NC GPIO kernel panic due to lack of register cleanup
This patch is ported from BSP R1.1 to do the following:

The NC GPIO driver does not properly tear down the required registers when
irq_free() are called causing the system to latch an IRQ which does not
exist any longer. The system gets a NULL-pointer dereference when it tries
to run an ISR that doesnt exist causing a kernel panic. The fix clears the
GPE, TPE, TNE and GTS registers for both core and SUS GPIOs when the IRQ is
disabled.

Original author is John Toomey <john.toomey@intel.com>

Signed-off-by: Tan, Raymond <raymond.tan@intel.com>
Change-Id: I71acf02d0462cff91d2d2956928ee22e7f411a44
Reviewed-on: https://git-gar-1.devtools.intel.com/gerrit/4992