x86: esram: Remove usage of macros that prevent deterministics build
In GCC 4.9, the usage of the __DATE__ and __TIME__ is be flagged as warnings
and treated as errors upon compilation. These 2 macros cause non
deterministics build with the value of each determined during preprocessing.

Since the usage is only for debug printout purpose, we have removed it instead
of altering the CFLAGS for compilation

Information:
https://gcc.gnu.org/onlinedocs/cpp/Standard-Predefined-Macros.html
https://gcc.gnu.org/gcc-4.9/changes.html

Signed-off-by: Tan, Raymond <raymond.tan@intel.com>
Change-Id: I441da9a8464783c2d9753b3fe4bd084234da3eb1
Reviewed-on: https://git-gar-1.devtools.intel.com/gerrit/4857
Reviewed-by: Ahmad, Josef <josef.ahmad@intel.com>