gpio-sch: fix irq handler bug (multiple synchrounous irqs) - RTC #75539
This patch is ported from R1.1 to do the following:

This patch fixes a bug that causes the driver to detect only a single
interrupt per power well at a time.  This bug is particularly nasty
as the undetected interrupts are then lost.

The root cause is a bug in the way interrupts are cleared.  The device
provides a single register that acts as both status and EOI (R/W1C).
The ISR in the driver is instead EOI'ing with a read-modify-write.
This causes any additional asserted bit to be cleared but never
served.

This fix is based on previous work by Billy Rafferty and Anthony Fee
(commits 8cac6f68 and 31ce6abd), and optimised to avoid code
duplication.

----

Signed-off-by: Tan, Raymond <raymond.tan@intel.com>
Signed-off-by: Arumugam,Kamenee <kamenee.arumugam@intel.com>
Change-Id: I87fb5c6dbea46371ad80e4931e413965014b9e52
Reviewed-on: https://git-gar-1.devtools.intel.com/gerrit/4994