platform: x86: Restrict probe mechanism for controlled probing
This patch is adding the restrict probe mechanism which was available in previous
release. While deferred probing has been enabled in kernel 3.14, but this
functionality is only available to both GPIO and I2C.

SPI on various quark platform has the dependency on GPIO, for its chip select, and the
function spi_match_master_to_boardinfo doesn't handle the EPROBE_DEFER, and if
SPI loads prior to the GPIO, the SPI device will fail to get initialized.

Signed-off-by: Tan, Raymond <raymond.tan@intel.com>
Change-Id: I780ed64eed3b3e8ee7ac1226b702c2a39b15c3a1
Reviewed-on: https://git-gar-1.devtools.intel.com/gerrit/5382
Reviewed-by: Fung, Joshua <joshua.fung@intel.com>