Revert "g-serial: CDC ACM enumeration as composite device - RTC #60147"
Previous fix solved device side problem did not implement a fix for PC driver.
To fix the PC side a new signed driver would have to be implmentented which
is not suitable in this product line.

Ultimately when fix is is necessary for a quark product and enforced on PC
strictly this fix can be implemented alongside the PC driver.

This reverts commit 68345df1c59d99a287de3285362f162295ff6cd1.

Signed-off-by: Kevin Clarke <kevin1.clarke@intel.com>
Change-Id: I0e64612d537fe6a7ecd1ddcda07b0b9ecb2e01b8
Reviewed-on: https://git-gar-1.devtools.intel.com/gerrit/5609
Tested-by: Tan, Raymond <raymond.tan@intel.com>
Reviewed-by: Fung, Joshua <joshua.fung@intel.com>