net: stmmac: fix a circular locking dependency
Change made by Zhang Xiao as part of Windriver kernel patch

netif_tx_lock spinlock as A, the tx_lock of stmmac's private
structure as B. In ndo_start_xmit process, A->B. While in function
stmmac_tx_clean, changed to B->A. In fact, it is not necessary
to hold the tx_lock for all the function stmmac_tx_clean.
Shrink its scope to avoid the conflict.

Signed-off-by: Zhang Xiao <xiao.zhang@windriver.com>
Signed-off-by: Kai Ji <kai.ji@intel.com>
Change-Id: I4b8a87be5d1be79d0ade21b88f087611274a7005
Reviewed-on: https://git-gar-1.devtools.intel.com/gerrit/5613
Reviewed-by: Tan, Raymond <raymond.tan@intel.com>
Reviewed-by: Fung, Joshua <joshua.fung@intel.com>