net: stmmac: fix kernel bug when releasing debugfs in stmmac. RTC# 114949
Change made by Zhang Jia as part of Windriver kernel patch

debugfs for stmmac driver is limited to be registered/unregistered one time
only because the registration/unregistration of debugfs for stmmac
driver is called along with adding/removing each interface. For
a multiple interface system, this counter should ensure the unregistration
of debugfs is called one time.

Signed-off-by: Lans Zhang <jia.zhang@windriver.com>
Signed-off-by: Kai Ji <kai.ji@intel.com>
Change-Id: I6b60dac748a925e0647172948d9b153c53712c29
Reviewed-on: https://git-gar-1.devtools.intel.com/gerrit/5617
Reviewed-by: Fung, Joshua <joshua.fung@intel.com>