platform: x86: intel_quark_platform_galileo: Change SPI, I2C & GPIO calling sequence (RTC:114524)
During the GPIO Restrict Probe on Galileo Gen1, the sequence of
calling SPI, I2C and GPIO interfaces is causing a system lock-up,
with "noapic nolapic" kernel parameter. The lock-up is due a
missing interrupt handler to handle the I2C interrupt (TX_EMPTY)
and it is because the GIP Probe would return without requesting
the IRQ when it fails to enable MSI.

The calling sequence is now changed to follow the Galileo Gen2
implementations so that the spi_register_board_info function
is called after I2C and GPIO during GPIO Restrict Probe to avoid
the system to lock-up.

Signed-off-by: Wong, Mun Choy <mun.choy.wong@intel.com>
Change-Id: Id29ffa2bb4bfd68676d70150661bf4d1ecf4033d
Reviewed-on: https://git-gar-1.devtools.intel.com/gerrit/5941
Reviewed-by: Tan, Raymond <raymond.tan@intel.com>
Tested-by: Tan, Raymond <raymond.tan@intel.com>